% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_rarecurve.R
\name{df_rarecurve}
\alias{df_rarecurve}
\title{df_rarecurve}
\usage{
df_rarecurve(comm)
}
\arguments{
\item{comm}{Community data, a matrix.}
}
\value{
A data_frame with the output of function \strong{rarefy}.
}
\description{
df_rarecurve
}
\examples{
require(vegan)
require(tidyverse)

data("dune")
data("BCI")

df_rarecurve(dune)

df_rarecurve(BCI) \%>\%
filter(sites == 10)

}
