% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_pair.R
\name{plot_beta_pair}
\alias{plot_beta_pair}
\title{plot_bet_pair}
\usage{
plot_beta_pair(x, index.family = "sorensen")
}
\arguments{
\item{x}{matriz de p/a o abundancia de especies.}

\item{index.family}{pude trabajar con "sorensen" o "jaccard".}
}
\value{
A plot.
}
\description{
Representacion grafica de las matrices de disimilaridad resultado de
la funcion "beta.pair". Visualizacion grafica de Turnover y Nestedness.
}
