% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rarecurve.R
\name{plot_rarecurve}
\alias{plot_rarecurve}
\title{plot_rarecurve}
\usage{
plot_rarecurve(comm)
}
\arguments{
\item{comm}{Community data, a matrix.}
}
\value{
A plot class \strong{ggplot}.
}
\description{
plot_rarecurve
}
\examples{
require(vegan)
data("dune")

plot_rarecurve(dune)
Manipulate the output with tidyverse.
plot_rarecurve(dune) +
  facet_wrap(.~sites) +
  theme(legend.position = "none")
}
