% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_specaccum.R
\name{plot_specaccum}
\alias{plot_specaccum}
\title{plot_specaccum}
\usage{
plot_specaccum(comm, method = "exact", permutations = 100)
}
\arguments{
\item{comm}{Community data set.}

\item{method}{The function use the same methods that specaccum function from
vegan package "random", "exact", "coleman", "rarefaction".
Except for "collector".}

\item{permutations}{Number of permutations with method = "random".}
}
\value{
A graph with ggplot2 characters.
}
\description{
plot_specaccum
}
\details{
For more info about the calculations checks help(specaccum).
}
\examples{
require(vegan)
data("BCI")

plot_specaccum(BCI)
plot_specaccum(BCI, method = "random", permutations = 1000)

Modify the layers:

plot_specaccum(BCI)+
theme_classic()+
theme(text = element_text(face = "bold"))

}
