% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_sample.R
\name{tidy.beta.sample}
\alias{tidy.beta.sample}
\title{tidy_beta_sample}
\usage{
tidy.beta.sample(x, index.family = "sorensen", sites = nrow(x$data),
  samples = 1)
}
\arguments{
\item{x}{matris de abundancia o presencia/ausencia.}

\item{index.family}{"jaccard" o "sorensen"}

\item{sites}{number of sites for which multiple-site dissimilarities will be computed. If not specified, default is all sites.}

\item{samples}{number of repetitions. If not specified, default is 1.}
}
\description{
tidy_beta_sample
}
