% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_multi.R
\name{tidy_beta_multi}
\alias{tidy_beta_multi}
\title{tidy_beta_multi}
\usage{
tidy_beta_multi(x, index.family = "sorensen")
}
\arguments{
\item{x}{matriz de laabundancia de especies.}

\item{index.family}{pude trabajar con "sorensen" o "jaccard".}
}
\value{
A a data_frame
}
\description{
tidy_beta_multi
}
