% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_multi_abund.R
\name{tidy_beta_multi_abund}
\alias{tidy_beta_multi_abund}
\title{tidy_beta_multi_abund}
\usage{
tidy_beta_multi_abund(x, index.family = "bray")
}
\arguments{
\item{x}{matriz de p/a o abundancia de especies.}

\item{index.family}{pude trabajar con "bray" o "ruzicka".}
}
\value{
A data_frame.
}
\description{
Representacion grafica de las matrices de disimilaridad resultado de
la funcion "beta.pair". Visualizacion grafica de Turnover y Nestedness.
}
