% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_pair.R
\name{tidy_beta_pair}
\alias{tidy_beta_pair}
\title{tidy_beta_pair}
\usage{
tidy_beta_pair(x, index.family = "sorensen")
}
\arguments{
\item{x}{matriz de p/a o abundancia de especies.}

\item{index.family}{pude trabajar con "jaccard" o "sorensen".}
}
\value{
A data_frame.
}
\description{
tidy_beta_pair
}
