% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_pair_abund.R
\name{tidy_beta_pair_abund}
\alias{tidy_beta_pair_abund}
\title{tidy_beta_pair_abund}
\usage{
tidy_beta_pair_abund(x, index.family = "bray")
}
\arguments{
\item{x}{matris de p/a o abundancia de especies.}

\item{index.family}{indices "bray" o "ruzicka"}
}
\value{
data_frame
}
\description{
tidy_beta_pair_abund
}
