% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_sample_abund.R
\name{tidy_beta_sample_abund}
\alias{tidy_beta_sample_abund}
\title{tidy_beta_sample_abund}
\usage{
tidy_beta_sample_abund(x, index.family = "bray", sites = nrow(x),
  samples = 1)
}
\arguments{
\item{x}{matris de abundancia de especies.}

\item{index.family}{"bray" o "ruzicka".}

\item{sites}{number of sites to be tested.}

\item{samples}{number of replicates.}
}
\description{
tidy_beta_sample_abund
}
