% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_beta_temp.R
\name{tidy_beta_temp}
\alias{tidy_beta_temp}
\title{tidy_beta_temp}
\usage{
tidy_beta_temp(x, y, index.family = "sorensen")
}
\arguments{
\item{x}{matris de abundancia inicial.}

\item{y}{matris de abundancia a contrastar.}

\item{index.family}{"jaccard" o "sorensen".}
}
\description{
tidy_beta_temp
}
