% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bray_part.R
\name{tidy_bray_part}
\alias{tidy_bray_part}
\title{tidy_bray_part}
\usage{
tidy_bray_part(x)
}
\arguments{
\item{x}{matris de abundancia de especies}
}
\value{
data_frame
}
\description{
tidy_bray_part
}
