% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functional_beta_multi.R
\name{tidy_functional_beta_multi}
\alias{tidy_functional_beta_multi}
\title{tidy_functional_beta_multi}
\usage{
tidy_functional_beta_multi(x, traits, index.family = "sorensen",
  warning.time = TRUE)
}
\arguments{
\item{x}{matris de abundancia.}

\item{traits}{traits info data.frame}

\item{index.family}{"jaccard" o "sorensen".}

\item{warning.time}{progres bar.}
}
\value{
data_frame
}
\description{
tidy_functional_beta_multi
}
