% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_functional_beta_pair.R
\name{tidy_functional_beta_pair}
\alias{tidy_functional_beta_pair}
\title{tidy_functional_beta_pair}
\usage{
tidy_functional_beta_pair(x, traits, index.family = "sorensen")
}
\arguments{
\item{x}{species matrix}

\item{traits}{traits dat.frame}

\item{index.family}{"jaccard" o "sorensen".}
}
\description{
tidy_functional_beta_pair
}
