% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_phylo_beta_multi.R
\name{tidy_phylo_beta_multi}
\alias{tidy_phylo_beta_multi}
\title{tidy_phylo_beta_multi}
\usage{
tidy_phylo_beta_multi(x, tree, index.family = "sorensen")
}
\arguments{
\item{x}{species matrix.}

\item{tree}{dendrogram}

\item{index.family}{"jaccard" or "sorensen".}
}
\value{
a data_frame.
}
\description{
tidy_phylo_beta_multi
}
