% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_phylo_beta_pair.R
\name{tidy_phylo_beta_pair}
\alias{tidy_phylo_beta_pair}
\title{tidy_phylo_beta_pair}
\usage{
tidy_phylo_beta_pair(x, tree, index.family = "sorensen")
}
\arguments{
\item{x}{matrix species.}

\item{tree}{dendrogram.}

\item{index.family}{"jaccard" or "sorensen".}
}
\value{
data_frame
}
\description{
tidy_phylo_beta_pair
}
